Split API screen into rough components
